requirements.txt
fastapi~=0.112.0
uvicorn~=0.30.5
SQLAlchemy~=2.0.32
python-dotenv~=1.0.1
mysql-connector-python~=9.0.0
celery[redis]~=5.4.0
pytest~=8.3.2
httpx~=0.27.0
pydantic~=2.8.2
isbnlib~=3.10.14
email_validator~=2.2.0
requests~=2.32.3
